+++ app/models/__init__.py
@@ -1,0 +1,8 @@
+from .assignment import Assignment
+from .question import Question
+from .batch_response import BatchResponse
+from .batch import Batch
+from .response import Response
+from .datatype import Datatype
+from .student import Student
+from .submission import Submission
