+++ app/models/student.py
@@ -1,0 +1,6 @@
+from app import db
+
+
+class Student(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    responses = db.relationship('Response', backref='student', lazy=True)
